+++ src/app/features/focus-mode/focus-mode-task-selection/focus-mode-task-selection.component.ts
@@ -24,7 +24,10 @@ export class FocusModeTaskSelectionComponent implements AfterViewInit, OnDestroy
   focusTimeout = 0;
   T: typeof T = T;
 
-  constructor(public readonly taskService: TaskService, private readonly _store: Store) {}
+  constructor(
+    public readonly taskService: TaskService,
+    private readonly _store: Store,
+  ) {}
 
   ngAfterViewInit(): void {
     this.focusTimeout = window.setTimeout(() => {
@@ -54,6 +57,7 @@ export class FocusModeTaskSelectionComponent implements AfterViewInit, OnDestroy
           setFocusSessionActivePage({ focusActivePage: FocusModePage.DurationSelection }),
         );
       } else {
+        this.taskService.setCurrentId(this.selectedTask.id);
         this._store.dispatch(
           setFocusSessionActivePage({ focusActivePage: FocusModePage.DurationSelection }),
         );