+++ src/app/features/tag/store/tag.reducer.ts
@@ -1,5 +1,4 @@
 import { createEntityAdapter, EntityAdapter } from '@ngrx/entity';
-import * as tagActions from './tag.actions';
 import { Tag, TagState } from '../tag.model';
 import { createFeatureSelector, createReducer, createSelector, on } from '@ngrx/store';
 import {
@@ -30,10 +29,21 @@ import {
 import { Update } from '@ngrx/entity/src/models';
 import { unique } from '../../../util/unique';
 import { loadAllData } from '../../../root-store/meta/load-all-data.action';
-import { TaskWithSubTasks } from '../../tasks/task.model';
 import { migrateTagState } from '../migrate-tag-state.util';
 import { MODEL_VERSION_KEY } from '../../../app.constants';
 import { MODEL_VERSION } from '../../../core/model-version';
+import {
+  addTag,
+  addToBreakTimeForTag,
+  deleteTag,
+  deleteTags,
+  updateAdvancedConfigForTag,
+  updateTag,
+  updateTagOrder,
+  updateWorkEndForTag,
+  updateWorkStartForTag,
+  upsertTag,
+} from './tag.actions';
 
 export const TAG_FEATURE_NAME = 'tag';
 const WORK_CONTEXT_TYPE: WorkContextType = WorkContextType.TAG;
@@ -218,23 +228,17 @@ export const tagReducer = createReducer<TagState>(
 
   // INTERNAL
   // --------
-  on(tagActions.addTag, (state: TagState, { tag }) => tagAdapter.addOne(tag, state)),
+  on(addTag, (state: TagState, { tag }) => tagAdapter.addOne(tag, state)),
 
-  on(tagActions.updateTag, (state: TagState, { tag }) =>
-    tagAdapter.updateOne(tag, state),
-  ),
+  on(updateTag, (state: TagState, { tag }) => tagAdapter.updateOne(tag, state)),
 
-  on(tagActions.upsertTag, (state: TagState, { tag }) =>
-    tagAdapter.upsertOne(tag, state),
-  ),
+  on(upsertTag, (state: TagState, { tag }) => tagAdapter.upsertOne(tag, state)),
 
-  on(tagActions.deleteTag, (state: TagState, { id }) => tagAdapter.removeOne(id, state)),
+  on(deleteTag, (state: TagState, { id }) => tagAdapter.removeOne(id, state)),
 
-  on(tagActions.deleteTags, (state: TagState, { ids }) =>
-    tagAdapter.removeMany(ids, state),
-  ),
+  on(deleteTags, (state: TagState, { ids }) => tagAdapter.removeMany(ids, state)),
 
-  on(tagActions.updateTagOrder, (state: TagState, { ids }) => {
+  on(updateTagOrder, (state: TagState, { ids }) => {
     if (ids.length !== state.ids.length) {
       console.log({ state, ids });
       throw new Error('Tag length should not change on re-order');
@@ -245,7 +249,7 @@ export const tagReducer = createReducer<TagState>(
     };
   }),
 
-  on(tagActions.updateWorkStartForTag, (state: TagState, { id, newVal, date }) =>
+  on(updateWorkStartForTag, (state: TagState, { id, newVal, date }) =>
     tagAdapter.updateOne(
       {
         id,
@@ -260,7 +264,7 @@ export const tagReducer = createReducer<TagState>(
     ),
   ),
 
-  on(tagActions.updateWorkEndForTag, (state: TagState, { id, newVal, date }) =>
+  on(updateWorkEndForTag, (state: TagState, { id, newVal, date }) =>
     tagAdapter.updateOne(
       {
         id,
@@ -275,7 +279,7 @@ export const tagReducer = createReducer<TagState>(
     ),
   ),
 
-  on(tagActions.addToBreakTimeForTag, (state: TagState, { id, valToAdd, date }) => {
+  on(addToBreakTimeForTag, (state: TagState, { id, valToAdd, date }) => {
     const oldTag = state.entities[id] as Tag;
     const oldBreakTime = oldTag.breakTime[date] || 0;
     const oldBreakNr = oldTag.breakNr[date] || 0;
@@ -297,27 +301,24 @@ export const tagReducer = createReducer<TagState>(
     );
   }),
 
-  on(
-    tagActions.updateAdvancedConfigForTag,
-    (state: TagState, { tagId, sectionKey, data }) => {
-      const tagToUpdate = state.entities[tagId] as Tag;
-      return tagAdapter.updateOne(
-        {
-          id: tagId,
-          changes: {
-            advancedCfg: {
-              ...tagToUpdate.advancedCfg,
-              [sectionKey]: {
-                ...tagToUpdate.advancedCfg[sectionKey],
-                ...data,
-              },
+  on(updateAdvancedConfigForTag, (state: TagState, { tagId, sectionKey, data }) => {
+    const tagToUpdate = state.entities[tagId] as Tag;
+    return tagAdapter.updateOne(
+      {
+        id: tagId,
+        changes: {
+          advancedCfg: {
+            ...tagToUpdate.advancedCfg,
+            [sectionKey]: {
+              ...tagToUpdate.advancedCfg[sectionKey],
+              ...data,
             },
           },
         },
-        state,
-      );
-    },
-  ),
+      },
+      state,
+    );
+  }),
 
   // TASK STUFF
   // ---------
@@ -344,11 +345,16 @@ export const tagReducer = createReducer<TagState>(
   }),
 
   on(deleteTask, (state, { task }) => {
-    const updates: Update<Tag>[] = task.tagIds.map((tagId) => ({
+    const affectedTagIds: string[] = [task, ...(task.subTasks || [])].reduce(
+      (acc, t) => [...acc, ...t.tagIds],
+      [] as string[],
+    );
+    const removedTasksIds: string[] = [task.id, ...(task.subTaskIds || [])];
+    const updates: Update<Tag>[] = affectedTagIds.map((tagId) => ({
       id: tagId,
       changes: {
         taskIds: (state.entities[tagId] as Tag).taskIds.filter(
-          (taskIdForTag) => taskIdForTag !== task.id,
+          (taskIdForTag) => !removedTasksIds.includes(taskIdForTag),
         ),
       },
     }));
@@ -356,9 +362,12 @@ export const tagReducer = createReducer<TagState>(
   }),
 
   on(moveToArchive_, (state, { tasks }) => {
-    const taskIdsToMoveToArchive = tasks.map((t) => t.id);
+    const taskIdsToMoveToArchive = tasks.flatMap((t) => [
+      t.id,
+      ...t.subTasks.map((st) => st.id),
+    ]);
     const tagIds = unique(
-      tasks.reduce((acc: string[], t: TaskWithSubTasks) => [...acc, ...t.tagIds], []),
+      tasks.flatMap((t) => [...t.tagIds, ...t.subTasks.flatMap((st) => st.tagIds)]),
     );
     const updates: Update<Tag>[] = tagIds.map((pid: string) => ({
       id: pid,
@@ -384,19 +393,31 @@ export const tagReducer = createReducer<TagState>(
     return tagAdapter.updateMany(updates, state);
   }),
 
-  on(restoreTask, (state, { task }) => {
-    return tagAdapter.updateMany(
-      task.tagIds
-        // NOTE: if the tag model is gone we don't update
-        .filter((tagId) => !!(state.entities[tagId] as Tag))
-        .map((tagId) => ({
-          id: tagId,
-          changes: {
-            taskIds: [...(state.entities[tagId] as Tag).taskIds, task.id],
-          },
-        })),
-      state,
-    );
+  on(restoreTask, (state, { task, subTasks }) => {
+    const allTasks = [task, ...subTasks];
+
+    // Create a map of tagIds to an array of associated task and subtask IDs
+    const tagTaskMap: { [tagId: string]: string[] } = {};
+    allTasks.forEach((t) => {
+      t.tagIds.forEach((tagId) => {
+        if (!tagTaskMap[tagId]) {
+          tagTaskMap[tagId] = [];
+        }
+        tagTaskMap[tagId].push(t.id);
+      });
+    });
+
+    // Create updates from the map
+    const updates = Object.entries(tagTaskMap)
+      .filter(([tagId]) => !!(state.entities[tagId] as Tag)) // If the tag model is gone we don't update
+      .map(([tagId, taskIds]) => ({
+        id: tagId,
+        changes: {
+          taskIds: [...(state.entities[tagId] as Tag).taskIds, ...taskIds],
+        },
+      }));
+
+    return tagAdapter.updateMany(updates, state);
   }),
 
   on(updateTaskTags, (state, { newTagIds = [], oldTagIds = [], task }) => {