+++ src/app/features/task-repeat-cfg/store/task-repeat-cfg.effects.ts
@@ -74,7 +74,7 @@ export class TaskRepeatCfgEffects {
       concatMap(
         () =>
           this._taskRepeatCfgService
-            .getRepeatTableTasksDueForDay$(
+            .getRepeatTableTasksDueForDayIncludingOverdue$(
               Date.now() - this._dateService.startOfNextDayDiff,
             )
             .pipe(first()),