+++ src/app/features/task-repeat-cfg/task-repeat-cfg.service.ts
@@ -4,7 +4,8 @@ import {
   selectAllTaskRepeatCfgs,
   selectTaskRepeatCfgById,
   selectTaskRepeatCfgByIdAllowUndefined,
-  selectTaskRepeatCfgsDueOnDay,
+  selectTaskRepeatCfgsDueOnDayIncludingOverdue,
+  selectTaskRepeatCfgsDueOnDayOnly,
   selectTaskRepeatCfgsWithStartTime,
 } from './store/task-repeat-cfg.reducer';
 import {
@@ -37,6 +38,7 @@ import { getDateTimeFromClockString } from '../../util/get-date-time-from-clock-
 import { isSameDay } from '../../util/is-same-day';
 import { remindOptionToMilliseconds } from '../tasks/util/remind-option-to-milliseconds';
 import { sortRepeatableTaskCfgs } from './sort-repeatable-task-cfg';
+import { getNewestPossibleDueDate } from './store/get-newest-possible-due-date.util';
 
 @Injectable({
   providedIn: 'root',
@@ -57,17 +59,24 @@ export class TaskRepeatCfgService {
     private _workContextService: WorkContextService,
   ) {}
 
-  getRepeatTableTasksDueForDay$(dayDate: number): Observable<TaskRepeatCfg[]> {
+  getRepeatTableTasksDueForDayOnly$(dayDate: number): Observable<TaskRepeatCfg[]> {
     // ===> taskRepeatCfgs scheduled for today and not yet created already
-    return this._store$.pipe(select(selectTaskRepeatCfgsDueOnDay, { dayDate }));
+    return this._store$.select(selectTaskRepeatCfgsDueOnDayOnly, { dayDate });
+  }
+
+  getRepeatTableTasksDueForDayIncludingOverdue$(
+    dayDate: number,
+  ): Observable<TaskRepeatCfg[]> {
+    // ===> taskRepeatCfgs scheduled for today and not yet created already
+    return this._store$.select(selectTaskRepeatCfgsDueOnDayIncludingOverdue, { dayDate });
   }
 
   getTaskRepeatCfgById$(id: string): Observable<TaskRepeatCfg> {
-    return this._store$.pipe(select(selectTaskRepeatCfgById, { id }));
+    return this._store$.select(selectTaskRepeatCfgById, { id });
   }
 
   getTaskRepeatCfgByIdAllowUndefined$(id: string): Observable<TaskRepeatCfg | undefined> {
-    return this._store$.pipe(select(selectTaskRepeatCfgByIdAllowUndefined, { id }));
+    return this._store$.select(selectTaskRepeatCfgByIdAllowUndefined, { id });
   }
 
   addTaskRepeatCfgToTask(
@@ -200,6 +209,13 @@ export class TaskRepeatCfgService {
     if (!isCreateNew) {
       return [];
     }
+    const targetCreated = getNewestPossibleDueDate(
+      taskRepeatCfg,
+      new Date(targetDayDate),
+    );
+    if (!targetCreated) {
+      throw new Error('Unable to getNewestPossibleDueDate()');
+    }
 
     const { task, isAddToBottom } = this._getTaskRepeatTemplate(taskRepeatCfg);
 
@@ -211,9 +227,9 @@ export class TaskRepeatCfgService {
       addTask({
         task: {
           ...task,
-          // NOTE otherwise isCreateNew check above would not work as intended and
+          // NOTE if moving this to top isCreateNew check above would not work as intended
           // we use created also for the repeat day label for past tasks
-          created: targetDayDate,
+          created: targetCreated.getTime(),
         },
         workContextType: this._workContextService
           .activeWorkContextType as WorkContextType,